requirements.txt
autogen
scholarly
python-dotenv
streamlit
requests
